# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Minecraft Operator API
  description: |
    REST API for managing PaperMC servers and plugins in Kubernetes.

    This API provides CRUD operations for PaperMCServer and Plugin custom resources,
    as well as action endpoints for triggering reconciliation and immediate updates.

    ## Authentication

    The API uses the same authentication as the Kubernetes cluster (service account tokens).

    ## Error Handling

    All errors return a JSON object with `error` field containing human-readable message,
    optional `code` for programmatic handling, and optional `details` for additional context.
  version: v1
  contact:
    name: Aleksei Sviridkin
    email: f@lex.la
  license:
    name: BSD-3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /servers:
    get:
      summary: List all PaperMC servers
      description: |
        Returns a list of all PaperMCServer resources in the cluster.
        Can be filtered by namespace using query parameter.
      operationId: listServers
      tags:
        - servers
      parameters:
        - name: namespace
          in: query
          description: Filter servers by Kubernetes namespace
          required: false
          schema:
            type: string
            example: minecraft
      responses:
        "200":
          description: Successfully retrieved list of servers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerListResponse"
              example:
                servers:
                  - name: survival-server
                    namespace: minecraft
                    currentVersion: "1.21.1-100"
                    desiredVersion: "1.21.1-105"
                    updateStrategy: auto
                    status: running
                    pluginCount: 5
                    availableUpdate:
                      version: "1.21.1"
                      build: 105
                      releasedAt: "2025-01-15T10:30:00Z"
                    nextMaintenance: "Every Sunday at 4:00"
                    labels:
                      environment: production
                      game-mode: survival
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new PaperMC server
      description: |
        Creates a new PaperMCServer custom resource in Kubernetes.
        The server will be automatically reconciled by the operator.
      operationId: createServer
      tags:
        - servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateRequest"
            example:
              name: creative-server
              namespace: minecraft
              updateStrategy: auto
              updateDelay: "168h"
              checkCron: "0 3 * * *"
              maintenanceWindow:
                enabled: true
                cron: "0 4 * * 0"
              gracefulShutdownTimeout: "5m"
              rcon:
                enabled: true
                passwordSecretName: creative-rcon-secret
                passwordSecretKey: password
                port: 25575
      responses:
        "201":
          description: Server created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /servers/{namespace}/{name}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/ServerName"

    get:
      summary: Get a specific server
      description: Returns detailed information about a specific PaperMCServer resource.
      operationId: getServer
      tags:
        - servers
      responses:
        "200":
          description: Server details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: Update a server
      description: |
        Updates an existing PaperMCServer resource.
        Only specified fields will be updated; unspecified fields remain unchanged.
      operationId: updateServer
      tags:
        - servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdateRequest"
      responses:
        "200":
          description: Server updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Delete a server
      description: |
        Deletes a PaperMCServer resource. This will also delete the associated
        StatefulSet, Service, and PVC (depending on reclaim policy).
      operationId: deleteServer
      tags:
        - servers
      responses:
        "204":
          description: Server deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      description: Kubernetes namespace
      schema:
        type: string
        minLength: 1
        maxLength: 63
        pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$"
        example: minecraft
    ServerName:
      name: name
      in: path
      required: true
      description: PaperMCServer resource name
      schema:
        type: string
        minLength: 1
        maxLength: 253
        pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
        example: survival-server
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Invalid request: missing required field 'name'"
            code: INVALID_REQUEST
            details:
              field: name
              reason: required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "PaperMCServer 'my-server' not found in namespace 'minecraft'"
            code: NOT_FOUND
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "PaperMCServer 'my-server' already exists in namespace 'minecraft'"
            code: ALREADY_EXISTS
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Failed to communicate with Kubernetes API"
            code: INTERNAL_ERROR
  schemas:
...
    UpdateStrategy:
      type: string
      description: |
        Defines how versions are managed for servers and plugins.

        - **latest**: Always use the latest available version from the repository
        - **auto**: Use constraint solver to find best version compatible with all dependencies
        - **pin**: Pin to specific version, auto-update to latest build within that version
        - **build-pin**: Pin to specific version AND build number, no automatic updates
      enum:
        - latest
        - auto
        - pin
        - build-pin
      example: auto
    ServerStatus:
      type: string
      description: |
        Current operational status of the server.

        - **running**: Server is running normally with all replicas ready
        - **updating**: Server is being updated (StatefulSet rolling update in progress)
        - **unknown**: Status cannot be determined (no condition or StatefulSet data)
      enum:
        - running
        - updating
        - unknown
      example: running
...
    ServiceType:
      type: string
      description: |
        Kubernetes Service type for the server.

        - **LoadBalancer**: External load balancer (requires cloud provider or MetalLB)
        - **NodePort**: Expose on a static port on each node
        - **ClusterIP**: Internal cluster IP only
      enum:
        - LoadBalancer
        - NodePort
        - ClusterIP
      default: LoadBalancer
      example: LoadBalancer
    ConditionStatus:
      type: string
      description: Status of a condition (True, False, or Unknown)
      enum:
        - "True"
        - "False"
        - Unknown
      example: "True"

    # ============================================================
    # SERVER SCHEMAS
    # ============================================================
    ServerListResponse:
      type: object
      required:
        - servers
      properties:
        servers:
          type: array
          description: List of server summaries
          items:
            $ref: "#/components/schemas/ServerSummary"
    ServerSummary:
      type: object
      description: Summary information about a PaperMCServer
      required:
        - name
        - namespace
        - updateStrategy
        - status
      properties:
        name:
          type: string
          description: Kubernetes resource name
          minLength: 1
          maxLength: 253
          example: survival-server
        namespace:
          type: string
          description: Kubernetes namespace
          example: minecraft
        currentVersion:
          type: string
          description: Currently running Paper version with build number (e.g., "1.21.1-100")
          example: "1.21.1-100"
        desiredVersion:
          type: string
          description: Target Paper version the operator wants to run
          example: "1.21.1-105"
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        status:
          $ref: "#/components/schemas/ServerStatus"
        pluginCount:
          type: integer
          description: Number of plugins matched to this server
          minimum: 0
          example: 5
        availableUpdate:
          $ref: "#/components/schemas/AvailableUpdate"
        nextMaintenance:
          type: string
          description: Human-readable description of next maintenance window
          example: "Every Sunday at 4:00"
        labels:
          type: object
          description: Kubernetes labels from the server resource
          additionalProperties:
            type: string
          example:
            environment: production
            game-mode: survival
    ServerDetail:
      type: object
      description: Detailed information about a PaperMCServer including plugins and schedules
      allOf:
        - $ref: "#/components/schemas/ServerSummary"
        - type: object
          properties:
            currentBuild:
              type: integer
              description: Currently running Paper build number
              minimum: 1
              example: 100
            desiredBuild:
              type: integer
              description: Target Paper build number
              minimum: 1
              example: 105
            plugins:
              type: array
              description: List of plugins matched to this server
              items:
                $ref: "#/components/schemas/ServerPlugin"
            maintenanceSchedule:
              $ref: "#/components/schemas/MaintenanceSchedule"
            lastUpdate:
              $ref: "#/components/schemas/UpdateHistory"
            updateBlocked:
              $ref: "#/components/schemas/UpdateBlockedStatus"
            conditions:
              type: array
              description: Kubernetes conditions for the server
              items:
                $ref: "#/components/schemas/Condition"
            rcon:
              $ref: "#/components/schemas/RCONStatus"
            service:
              $ref: "#/components/schemas/ServiceStatus"
    ServerPlugin:
      type: object
      description: Status of a plugin on a specific server
      required:
        - name
        - compatible
      properties:
        name:
          type: string
          description: Plugin resource name
          example: bluemap
        namespace:
          type: string
          description: Plugin resource namespace
          example: minecraft
        project:
          type: string
          description: Project identifier in the repository
          example: BlueMap
        source:
          type: string
          description: Plugin source type
          example: hangar
        resolvedVersion:
          type: string
          description: Version resolved by the constraint solver
          example: "5.4"
        currentVersion:
          type: string
          description: Currently installed version on the server
          example: "5.3"
        desiredVersion:
          type: string
          description: Target version to install
          example: "5.4"
        compatible:
          type: boolean
          description: Whether the plugin is compatible with the server version
          example: true
        pendingDeletion:
          type: boolean
          description: Whether the plugin is marked for removal
          default: false
          example: false
        installedJarName:
          type: string
          description: Filename of the installed JAR in /data/plugins/
          example: "BlueMap-5.4-paper.jar"
    ServerCreateRequest:
      type: object
      description: Request body for creating a new PaperMCServer
      required:
        - name
        - namespace
        - updateStrategy
      properties:
        name:
          type: string
          description: Kubernetes resource name (must be valid DNS subdomain)
          minLength: 1
          maxLength: 253
          pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
          example: creative-server
        namespace:
          type: string
          description: Kubernetes namespace to create the server in
          example: minecraft
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        version:
          type: string
          description: |
            Target Paper version (required for pin and build-pin strategies).
            Example: "1.21.1" for Minecraft 1.21.1
          pattern: "^\\d+\\.\\d+(\\.\\d+)?$"
          example: "1.21.1"
        build:
          type: integer
          description: |
            Target Paper build number (required for build-pin strategy).
            For pin strategy, serves as minimum build.
          minimum: 1
          example: 100
        updateDelay:
          type: string
          description: |
            Grace period before applying updates (Go duration format).
            Examples: "168h" (7 days), "24h", "30m"
          pattern: "^(\\d+h)?(\\d+m)?(\\d+s)?$"
          example: "168h"
        checkCron:
          type: string
          description: |
            Cron expression for update checks.
            Format: minute hour day month weekday
          example: "0 3 * * *"
        maintenanceWindow:
          $ref: "#/components/schemas/MaintenanceWindowConfig"
        gracefulShutdownTimeout:
          type: string
          description: |
            Graceful shutdown timeout (Go duration format).
            Must match StatefulSet terminationGracePeriodSeconds.
          default: "5m"
          example: "5m"
        rcon:
          $ref: "#/components/schemas/RCONConfig"
        service:
          $ref: "#/components/schemas/ServiceConfig"
        labels:
          type: object
          description: Labels to apply to the server resource
          additionalProperties:
            type: string
          example:
            environment: production
            game-mode: creative
    ServerUpdateRequest:
      type: object
      description: Request body for updating an existing PaperMCServer
      properties:
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        version:
          type: string
          description: Target Paper version
          example: "1.21.1"
        build:
          type: integer
          description: Target Paper build number
          minimum: 1
        updateDelay:
          type: string
          description: Grace period before applying updates
          example: "168h"
        checkCron:
          type: string
          description: Cron expression for update checks
          example: "0 3 * * *"
        maintenanceWindow:
          $ref: "#/components/schemas/MaintenanceWindowConfig"
        labels:
          type: object
          description: Labels to set on the server resource
          additionalProperties:
            type: string

    # ============================================================
    # PLUGIN SCHEMAS
    # ============================================================
...
    AvailableUpdate:
      type: object
      description: Information about an available update
      properties:
        version:
          type: string
          description: Available Paper version
          example: "1.21.1"
        build:
          type: integer
          description: Available Paper build number
          minimum: 1
          example: 105
        releasedAt:
          type: string
          format: date-time
          description: When this version was released
          example: "2025-01-15T10:30:00Z"
        foundAt:
          type: string
          format: date-time
          description: When this update was discovered
          example: "2025-01-15T11:00:00Z"
        plugins:
          type: array
          description: Plugin versions for this update
          items:
            $ref: "#/components/schemas/PluginVersionPair"
...
    MaintenanceSchedule:
      type: object
      description: Maintenance schedule configuration
      properties:
        checkCron:
          type: string
          description: Cron expression for update checks
          example: "0 3 * * *"
        windowCron:
          type: string
          description: Cron expression for maintenance window
          example: "0 4 * * 0"
        nextWindow:
          type: string
          description: Human-readable next maintenance window
          example: "Every Sunday at 4:00"
        enabled:
          type: boolean
          description: Whether maintenance window is enabled
          example: true
    MaintenanceWindowConfig:
      type: object
      description: Maintenance window configuration for create/update requests
      properties:
        enabled:
          type: boolean
          description: Whether to enable maintenance window
          default: false
          example: true
        cron:
          type: string
          description: Cron expression for maintenance window
          example: "0 4 * * 0"
    UpdateHistory:
      type: object
      description: Record of the last update attempt
      properties:
        appliedAt:
          type: string
          format: date-time
          description: When the update was applied
          example: "2025-01-14T04:00:00Z"
        previousVersion:
          type: string
          description: Version before the update
          example: "1.21.1-95"
        successful:
          type: boolean
          description: Whether the update succeeded
          example: true
    UpdateBlockedStatus:
      type: object
      description: Information about blocked updates
      properties:
        blocked:
          type: boolean
          description: Whether updates are currently blocked
          example: true
        reason:
          type: string
          description: Human-readable reason for the block
          example: "Plugin 'bluemap' is incompatible with target version"
        blockedBy:
          $ref: "#/components/schemas/BlockedByInfo"
    BlockedByInfo:
      type: object
      description: Details about what is blocking an update
      properties:
        plugin:
          type: string
          description: Name of the blocking plugin
          example: bluemap
        version:
          type: string
          description: Version of the blocking plugin
          example: "5.3"
        supportedVersions:
          type: array
          description: Minecraft versions this plugin supports
          items:
            type: string
          example:
            - "1.20.4"
            - "1.21"
    RCONConfig:
      type: object
      description: RCON configuration for create/update requests
      properties:
        enabled:
          type: boolean
          description: Whether RCON is enabled
          default: false
          example: true
        passwordSecretName:
          type: string
          description: Name of the Secret containing the RCON password
          example: my-rcon-secret
        passwordSecretKey:
          type: string
          description: Key within the Secret containing the password
          default: password
          example: password
        port:
          type: integer
          description: RCON port
          minimum: 1
          maximum: 65535
          default: 25575
          example: 25575
    RCONStatus:
      type: object
      description: RCON status information
      properties:
        enabled:
          type: boolean
          description: Whether RCON is enabled
          example: true
        port:
          type: integer
          description: RCON port
          example: 25575
    ServiceConfig:
      type: object
      description: Kubernetes Service configuration
      properties:
        type:
          $ref: "#/components/schemas/ServiceType"
        annotations:
          type: object
          description: Custom annotations for the Service
          additionalProperties:
            type: string
          example:
            metallb.universe.tf/loadBalancerIPs: "192.168.1.100"
        loadBalancerIP:
          type: string
          description: Static IP for LoadBalancer type
          format: ipv4
          example: "192.168.1.100"
    ServiceStatus:
      type: object
      description: Service status information
      properties:
        type:
          $ref: "#/components/schemas/ServiceType"
        clusterIP:
          type: string
          description: Cluster IP address
          example: "10.96.100.50"
        externalIP:
          type: string
          description: External IP address (for LoadBalancer)
          example: "192.168.1.100"
...
    Condition:
      type: object
      description: Kubernetes condition
      required:
        - type
        - status
      properties:
        type:
          type: string
          description: Condition type
          example: Ready
        status:
          $ref: "#/components/schemas/ConditionStatus"
        reason:
          type: string
          description: Machine-readable reason for the condition
          example: ReconciliationSucceeded
        message:
          type: string
          description: Human-readable message
          example: "Server is running and ready"
        lastTransitionTime:
          type: string
          format: date-time
          description: When the condition last changed
          example: "2025-01-15T10:30:00Z"
        observedGeneration:
          type: integer
          description: Generation of the resource when condition was set
          format: int64
          example: 5
...
    ErrorResponse:
      type: object
      description: Error response body
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Failed to get server: not found"
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - NOT_FOUND
            - ALREADY_EXISTS
            - INTERNAL_ERROR
            - FORBIDDEN
            - UNAUTHORIZED
          example: NOT_FOUND
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          example:
            field: name
            reason: invalid format